Enable windows support (#121)
* manual merge complete for the patching task - phew!

* I think that is all of the manual rebasing

* Fix up typos from merge

* Ensure pdk validate and pdk test unit work on all new platforms

* Fix makefile ordering, bad depends and some limus params

* cleaned up code since split to two scripts

* re-format

* remove refreshfact parameter

* remove unneeded "wsus" stuff that didn't work

* mod for windows

* update windows versions

* fix windows sch task

* correct windows path

* missing $

* make is_windows constant

* disable unneeded pre-scan on windows, fix comment

* Rubocop fixen

* Final updates for new version of litmus

* Turns out apply_idempotent does two applies

* Rev version number

* Updated changelog for V0.11.0

* Missed Jake off the README last time

+++ spec/classes/os_patching_spec.rb
@@ -5,6 +5,15 @@ describe 'os_patching' do
     context "on #{os}" do
       let(:facts) { os_facts }
 
+      case os_facts[:kernel]
+      when 'Linux'
+        let(:cache_dir) { '/var/cache/os_patching' }
+        let(:fact_cmd) { '/usr/local/bin/os_patching_fact_generation.sh' }
+      when 'windows'
+        let(:cache_dir) { 'C:/ProgramData/os_patching' }
+        let(:fact_cmd) { 'C:/ProgramData/os_patching/os_patching_fact_generation.ps1' }
+      end
+
       case os_facts[:osfamily]
       when 'RedHat'
         context 'with package management enabled' do
@@ -28,24 +37,18 @@ describe 'os_patching' do
 
       context 'with reboot_override => always' do
         let(:params) { {'reboot_override' => 'always'} }
-        it { is_expected.to contain_file('/var/cache/os_patching/reboot_override').with({
+        it { is_expected.to contain_file("#{cache_dir}/reboot_override").with({
           'ensure' => 'file',
-          'owner'  => 'root',
-          'group'  => 'root',
-          'mode'   => '0644',
         })}
-        it { is_expected.to contain_file('/var/cache/os_patching/reboot_override').with_content(/^always$/)}
+        it { is_expected.to contain_file("#{cache_dir}/reboot_override").with_content(/^always$/)}
       end
 
       context 'with reboot_override => never' do
         let(:params) { {'reboot_override' => 'never'} }
-        it { is_expected.to contain_file('/var/cache/os_patching/reboot_override').with({
+        it { is_expected.to contain_file("#{cache_dir}/reboot_override").with({
           'ensure' => 'file',
-          'owner'  => 'root',
-          'group'  => 'root',
-          'mode'   => '0644',
         })}
-        it { is_expected.to contain_file('/var/cache/os_patching/reboot_override').with_content(/^never$/)}
+        it { is_expected.to contain_file("#{cache_dir}/reboot_override").with_content(/^never$/)}
       end
 
       context 'with reboot_override => foobar' do
@@ -60,13 +63,10 @@ describe 'os_patching' do
 
       context 'with patch_window => Week3' do
         let(:params) { {'patch_window' => 'Week3'} }
-        it { is_expected.to contain_file('/var/cache/os_patching/patch_window').with({
+        it { is_expected.to contain_file(cache_dir + '/patch_window').with({
           'ensure' => 'file',
-          'owner'  => 'root',
-          'group'  => 'root',
-          'mode'   => '0644',
         })}
-        it { is_expected.to contain_file('/var/cache/os_patching/patch_window').with_content(/^Week3$/)}
+        it { is_expected.to contain_file(cache_dir + '/patch_window').with_content(/^Week3$/)}
       end
 
       context 'with blackout window set' do
@@ -75,57 +75,48 @@ describe 'os_patching' do
             'blackout_windows' => { 'End of year change freeze': { 'start': '2018-12-15T00:00:00+10:00', 'end': '2019-01-15T23:59:59+10:00' } }
           }
         }
-        it { is_expected.to contain_file('/var/cache/os_patching/blackout_windows').with({
+        it { is_expected.to contain_file(cache_dir + '/blackout_windows').with({
           'ensure' => 'file',
-          'owner'  => 'root',
-          'group'  => 'root',
-          'mode'   => '0644',
         })}
-        it { is_expected.to contain_file('/var/cache/os_patching/blackout_windows').with_content(/End of year change/)}
+        it { is_expected.to contain_file(cache_dir + '/blackout_windows').with_content(/End of year change/)}
       end
 
       it { is_expected.to compile }
       it { is_expected.to compile.with_all_deps }
       it { is_expected.to contain_class('os_patching') }
-      it { is_expected.to contain_file('/var/cache/os_patching').with({
+      it { is_expected.to contain_file(cache_dir).with({
         'ensure' => 'directory',
-        'owner'  => 'root',
-        'group'  => 'root',
-        'mode'   => '0644',
       })}
 
-      it { is_expected.to contain_file('/var/cache/os_patching/blackout_windows').with({
+      it { is_expected.to contain_file(cache_dir + '/blackout_windows').with({
         'ensure' => 'absent',
       })}
 
-      it { is_expected.to contain_file('/var/cache/os_patching/patch_window').with({
+      it { is_expected.to contain_file(cache_dir + '/patch_window').with({
         'ensure' => 'absent',
       })}
 
-      it { is_expected.to contain_file('/var/cache/os_patching/reboot_override').with({
+      it { is_expected.to contain_file(cache_dir + '/reboot_override').with({
         'ensure' => 'file',
-        'owner'  => 'root',
-        'group'  => 'root',
-        'mode'   => '0644',
       })}
 
-      it { is_expected.to contain_file('/var/cache/os_patching/reboot_override').with_content(/^default$/)}
+      it { is_expected.to contain_file(cache_dir + '/reboot_override').with_content(/^default$/)}
 
-      it { is_expected.to contain_file('/usr/local/bin/os_patching_fact_generation.sh').with({
+      it { is_expected.to contain_file(fact_cmd).with({
         'ensure' => 'file',
-        'owner'  => 'root',
-        'group'  => 'root',
-        'mode'   => '0700',
       })}
 
-      it { is_expected.to contain_cron('Cache patching data').with_ensure('present') }
-      it { is_expected.to contain_cron('Cache patching data at reboot').with_ensure('present') }
+      case os_facts[:kernel]
+      when 'Linux'
+        it { is_expected.to contain_cron('Cache patching data').with_ensure('present') }
+        it { is_expected.to contain_cron('Cache patching data at reboot').with_ensure('present') }
+      end
       it { is_expected.to contain_exec('os_patching::exec::fact') }
       it { is_expected.to contain_exec('os_patching::exec::fact_upload') }
 
       context 'purge module' do
         let(:params) { {'ensure' => 'absent'} }
-        it { is_expected.to contain_file('/var/cache/os_patching').with({
+        it { is_expected.to contain_file(cache_dir).with({
           'ensure' => 'absent',
         })}
       end
